ranges

+++ SHIRPI/templatetags/comment_filters.py
@@ -14,7 +14,8 @@ def display_comment(value, user,autoescape=None):
 	result = "<h3 class='timestamp'>" 
 	result += "<a href='#" + urllib.quote_plus(str(value.id)) + "' name ='" + urllib.quote_plus(str(value.id)) + "'>" + str(value.last_modified.ctime()) + "</a>"	#Anchor position can be better set when layout is more mature
 	result += "</h3>\n"
-	result += "<h4 class='delete'><a href='/cs215/shirpi/delete_comment/" + str(value.id) + "'>Delete</a></h4>"
+	if user.username == value.author.username or user.has_perm("SHIRPI.comment"):
+		result += "<h4 class='delete'><a href='/cs215/shirpi/delete_comment/" + str(value.id) + "'>Delete</a></h4>"
 	result += "<h3 class='username'>"
 	if value.author.username != "Anonymous":
 	        result += "<a href='/cs215/shirpi/view_profile/" + urllib.quote_plus(value.author.username) + "/'>"
